third_party/blink/web_tests/platform/generic/fast/forms/range/range-hit-test-with-padding-expected.txt

Clicking middle of input type=range with padding should set the value to middle.

PASS input_ltr.value is "25"
PASS input_rtl.value is "25"
PASS input_vertical_ltr.value is "25"
PASS input_vertical_rtl.value is "25"
PASS successfullyParsed is true

TEST COMPLETE

   
